Finished Section B2 and B3
For time constraints, I've removed B1 from the project and will need to revisit at a later time.  

B1 involves separating out the "Attributes" dimension, which is a list within a value.  Would need to be able to have listing_id with the attribute list, then create a binary classifier on [if listing has attribute, 1, else 0].  Then it would be possible to do attribute specific analysis.